Fix #3. Convert source files with double spaces in their filename.
Previously, the default internal field separator (IFS) was being used. This is "<space><tab><newline>" and 2 or more repeated characters was reduce to a single separator during substitutions. To fix this, IFS is now cleared before making any substitutions that involve filenames. (Filenames can contain any character so there are no safe characters to use as the IFS).